Created scales helper and removed another duplicate setDimensions
