+++ src/helpers/helpers.plugins.js
@@ -1,7 +1,7 @@
 'use strict';
 
 /**
- * @namespace Chart.helpers.controllers
+ * @namespace Chart.helpers.plugins
  */
 module.exports = {
 	/**
@@ -35,36 +35,5 @@ module.exports = {
 		me.afterUpdate();
 
 		return me.minSize;
-	},
-	/**
-	 * Sets dimensions for plugin
-	 */
-	setDimensions: function() {
-		var me = this;
-		// Set the unconstrained dimension before label rotation
-		if (me.isHorizontal()) {
-			// Reset position before calculating rotation
-			me.width = me.maxWidth;
-			me.left = 0;
-			me.right = me.width;
-		} else {
-			me.height = me.maxHeight;
-
-			// Reset position before calculating rotation
-			me.top = 0;
-			me.bottom = me.height;
-		}
-
-		// Reset padding
-		me.paddingLeft = 0;
-		me.paddingTop = 0;
-		me.paddingRight = 0;
-		me.paddingBottom = 0;
-
-		// Reset minSize
-		me.minSize = {
-			width: 0,
-			height: 0
-		};
 	}
 };
