+++ src/helpers/index.js
@@ -6,3 +6,4 @@ module.exports.canvas = require('./helpers.canvas');
 module.exports.options = require('./helpers.options');
 module.exports.controllers = require('./helpers.controllers');
 module.exports.plugins = require('./helpers.plugins');
+module.exports.scale = require('./helpers.scale');
