+++ src/plugins/plugin.legend.js
@@ -119,7 +119,7 @@ module.exports = function(Chart) {
 		//
 
 		beforeSetDimensions: noop,
-		setDimensions: helpers.plugins.setDimensions,
+		setDimensions: helpers.scale.setDimensions,
 		afterSetDimensions: noop,
 
 		//
